refactor: move memoize
